LayoutTests/inspector/profiler/webgl/webgl-profiler-api-changes-expected.txt

Test to catch WebGL API changes. If this test should ever fail, we should re-examine the WebGL state saving/restoring logic in the InjectedScriptModule to include any latest changes to the API.

Constants for gl.isEnabled():
BLEND
CULL_FACE
DEPTH_TEST
DITHER
POLYGON_OFFSET_FILL
SAMPLE_ALPHA_TO_COVERAGE
SAMPLE_COVERAGE
SCISSOR_TEST
STENCIL_TEST

Constants for gl.pixelStorei():
PACK_ALIGNMENT
UNPACK_ALIGNMENT
UNPACK_COLORSPACE_CONVERSION_WEBGL
UNPACK_FLIP_Y_WEBGL
UNPACK_PREMULTIPLY_ALPHA_WEBGL

Constants for gl.getTexParameter():
TEXTURE_MAG_FILTER
TEXTURE_MIN_FILTER
TEXTURE_WRAP_S
TEXTURE_WRAP_T

Constants for gl.getVertexAttrib():
CURRENT_VERTEX_ATTRIB
VERTEX_ATTRIB_ARRAY_BUFFER_BINDING
VERTEX_ATTRIB_ARRAY_ENABLED
VERTEX_ATTRIB_ARRAY_NORMALIZED
VERTEX_ATTRIB_ARRAY_SIZE
VERTEX_ATTRIB_ARRAY_STRIDE
VERTEX_ATTRIB_ARRAY_TYPE

Constants for gl.getVertexAttribOffset():
VERTEX_ATTRIB_ARRAY_POINTER

Constants for gl.getParameter() that we do not track in InjectedScriptModule:
ALIASED_LINE_WIDTH_RANGE
ALIASED_POINT_SIZE_RANGE
ALPHA_BITS
BLUE_BITS
COMPRESSED_TEXTURE_FORMATS
DEPTH_BITS
GREEN_BITS
MAX_COMBINED_TEXTURE_IMAGE_UNITS
MAX_CUBE_MAP_TEXTURE_SIZE
MAX_FRAGMENT_UNIFORM_VECTORS
MAX_RENDERBUFFER_SIZE
MAX_TEXTURE_IMAGE_UNITS
MAX_TEXTURE_SIZE
MAX_VARYING_VECTORS
MAX_VERTEX_ATTRIBS
MAX_VERTEX_TEXTURE_IMAGE_UNITS
MAX_VERTEX_UNIFORM_VECTORS
MAX_VIEWPORT_DIMS
RED_BITS
RENDERER
SAMPLES
SAMPLE_BUFFERS
SHADING_LANGUAGE_VERSION
STENCIL_BITS
SUBPIXEL_BITS
VENDOR
VERSION
